Testing infrastructure with self-hosted runner (#3)
By using a self-hosted runner it is possible to leave the kind cluster running and reconcile/retest branch changes quickly. PRs will be build from scratch.

Running multiple builds/tests in parallel is probably not working yet, this approach is still WIP.

Also this is a step forward to run (and keep running) the management cluster (flux & cluster-api) in kind for "real world clouds" testing.