Move admission controllers to extension meta chart
Admission controllers are enabled by specifying admission=true in
extension-base-values.yaml now. Consequently, the amount of new files
is reduced.

Don't enable the hlcoud admission controller, as it is not functional
yet and shoot creation will fail in consequence. Once, a functional
admission controller exists, it can be enabled, too.

Signed-off-by: Jens Schneider <schneider@23technologies.cloud>