Add pre-push hook.
This should make sure that the internal chart versions match the tag,
which should be pushed. As this is hard to test in isolation, we
should simple collect some experience with it.

Also, we should not create any tags via the github webinterface, in
order to keep the repository (and especially the releases) clean.

Signed-off-by: Jens Schneider <schneider@23technologies.cloud>