Fail the pipeline when image tag includes latest (#734)
* Add test searching for image tags including latest

We do not want to roll out images which are referring to a latest
tag. Therefore, let the pipeline fail in that case. This is a very
late point in the overall process to catch such an error, however, it
prevents us from merging erroneous updates in case the gardener chart
releaser was not able to set the correct image tag for some reason.

---------
Signed-off-by: Jens Schneider <schneider@23technologies.cloud>