Add input option for version tag
This update introduces an optional user input to the build-and-push workflow, used to specify the version tag of gardener/gardener. This promotes flexibility and user control over the version used in the build. Additionally, the workflow checks for the existence of the specified remote image before initiating the build, thereby preventing unnecessary builds.

Signed-off-by: Jan Lohage <lohage@23technologies.cloud>